logging does no longer fail in platform init
