+++ src/api/nabto_device_config.h
@@ -1,0 +1,8 @@
+#ifndef _NABTO_DEVICE_CONFIG_H_
+#define _NABTO_DEVICE_CONFIG_H_
+
+#ifndef NABTO_DEVICE_LOG_STD_OUT_CALLBACK
+#define NABTO_DEVICE_LOG_STD_OUT_CALLBACK 1
+#endif
+
+#endif
