+++ src/api/nabto_device_logging_std_out_callback.h
@@ -1,0 +1,8 @@
+#ifndef _NABTO_DEVICE_STD_OUT_LOGGER_
+#define _NABTO_DEVICE_STD_OUT_LOGGER_
+
+#include <nabto/nabto_device.h>
+
+void nabto_device_logging_std_out_callback(NabtoDeviceLogMessage* msg, void* data);
+
+#endif
