+++ src/modules/logging/api/CMakeLists.txt
@@ -1,0 +1,3 @@
+set(src std_out_logger.c)
+
+add_library(nm_logging_api "${src}")
