+++ src/nabto_device_libevent/CMakeLists.txt
@@ -27,6 +27,7 @@ target_link_libraries(nabto_device_static
   nm_mbedtls_random
   nm_mdns
   nm_tcp_tunnel
+  nm_logging_api
   nabto_device_threads
   3rdparty_mbedtls
   )
