+++ src/nabto_device_libevent/nabto_platform_libevent.c
@@ -19,6 +19,11 @@
 
 #include <stdlib.h>
 
+#include <platform/np_logging.h>
+#include <platform/np_logging_defines.h>
+
+#define LOG NABTO_LOG_MODULE_PLATFORM
+
 static void signal_event(evutil_socket_t s, short event, void* userData);
 static void* libevent_thread(void* data);
 
@@ -44,6 +49,7 @@ struct libevent_platform {
  */
 np_error_code nabto_device_platform_init(struct nabto_device_context* device, struct nabto_device_mutex* eventMutex)
 {
+    NABTO_LOG_TRACE(LOG, "initializing platform");
     // Initialize the global libevent context.
     nm_libevent_global_init();
 
@@ -51,7 +57,6 @@ np_error_code nabto_device_platform_init(struct nabto_device_context* device, st
     // functionality which can change between the nabto_device
     // implementations.
     struct libevent_platform* platform = calloc(1, sizeof(struct libevent_platform));
-
 
     platform->eventBase = event_base_new();
     platform->signalEvent = event_new(platform->eventBase, -1, 0, &signal_event, platform);
